requirements.txt
Flask==3.1.0
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0
click==8.2.1
blinker==1.9.0
gunicorn==20.1.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.40
Flask-Mail==0.9.1
Flask-WTF==1.1.1
WTForms==3.2.1
email-validator==2.0.0
PyMySQL==1.1.1
SpeechRecognition==3.10.0
pydub==0.25.1
nltk==3.8.1
numpy==1.24.3
fpdf2==2.7.4
reportlab==4.0.4
requests==2.32.3
